+++ app/templates/api/responses/notFound.js
@@ -6,11 +6,11 @@
  * Used when the requested resource is not found, whether it doesn't exist.
  */
 
-module.exports = function (data, status, message) {
+module.exports = function (data, code, message, root) {
     this.res.status(404);
-    this.res.jsonx({
-        status: status || "E_NOT_FOUND",
-        message: message || "Requested resource is not found",
+    this.res.jsonx(_.assign({
+        code: code || "E_NOT_FOUND",
+        message: message || "The requested resource could not be found but may be available again in the future",
         response: data || {}
-    });
+    }, root));
 };