+++ app/templates/api/responses/serverError.js
@@ -1,15 +1,15 @@
 /**
- * 500 (Server Error) Response
+ * 500 (Internal Server Error) Response
  *
  * A generic error message, given when no more specific message is suitable.
  * The general catch-all error when the server-side throws an exception.
  */
 
-module.exports = function (data, status, message) {
+module.exports = function (data, code, message, root) {
     this.res.status(500);
-    this.res.jsonx({
-        status: status || "E_SERVER_ERROR",
-        message: message || "Internal server error",
+    this.res.jsonx(_.assign({
+        code: code || "E_INTERNAL_SERVER_ERROR",
+        message: message || "Something bad happened on the server",
         response: data || {}
-    });
+    }, root));
 };