+++ app/templates/api/responses/unauthorized.js
@@ -6,6 +6,11 @@
  * Error code response for missing or invalid authentication token.
  */
 
-module.exports = function () {
+module.exports = function (data, code, message, root) {
     this.res.status(401);
+    this.res.jsonx(_.assign({
+        code: code || "E_UNAUTHORIZED",
+        message: message || "Missing or invalid authentication token",
+        response: data || {}
+    }, root));
 };